Change from `Panel` to `PANEL`
Previously I have a plan to make comment API named as `window.Comment`.
But, apparently Firefox has a `window.Comment` object already. So I have
to choose another name alternative and I think the most semantic and
easiest name is `window.COMMENT`.

We also have `window.Panel` object previously, and now I have to change
the object name to `window.PANEL` for consistency with the comment API
in the future.