Feature
Added ability to activate a tab and set pre-defined field values through query string URL. This is useful for cases if you want somebody to fill a form like this and that but at the same time you don’t want to say much. Just paste the URL! Make sure to encode characters like `:` and `+`, just in case.
git commit -m Feature