Move Pager to the First Tab
So that switching tab when there is pager visible will not preserve the
pager visibility on other tabs. I hope you understand what I mean.

Unlike the action buttons in page editor for example, they are available
on the third section of the form. So, switching between tabs will not
change the action buttons visibility.

Previously, pager element was placed in the third section of the form,
just like the action buttons.