Prefer to Use `.index.php` to Disable Extension
And layout as well. So that other layout can force to activate the
disabled layout by requiring that `.index.php` file to the current
layout's `index.php` file. For example, to make a child theme (layout)
that depends on parent layout. But because the parent layout currently
disabled, author need to activate it manually by requiring it to the
child theme's (layout's) `index.php` file.